requirements.txt
pandas==1.4.4
plotly
psycopg2-binary
pydeck
pymongo[srv]
python-dotenv
SQLAlchemy
streamlit
